feat(cms): removes anonym objects from cms schemas
Required by Sanity in order to support GraphQL
