--- cms/src/schemas/custom/year.ts
@@ -1,6 +1,0 @@
-export default {
-  name: "year",
-  title: "Year",
-  type: "number",
-  validation: (Rule) => [Rule.required(), Rule.integer(), Rule.positive()],
-}
