+++ cms/src/schemas/objects/media.ts
