+++ cms/src/schemas/objects/mediaImage.ts
