+++ cms/src/schemas/objects/settingsCatchphrases.ts
@@ -1,0 +1,51 @@
+export default {
+  name: "settingsCatchphrase",
+  title: "Catchphrase",
+  type: "object",
+  fields: [
+    {
+      name: "text",
+      title: "Text",
+      type: "blockRichText",
+    },
+    {
+      name: "visibility",
+      title: "Visibility",
+      type: "string",
+      validation: (Rule) => Rule.required(),
+      initialValue: "BOTH",
+      options: {
+        layout: "radio",
+        list: [
+          { title: "Desktop and mobile", value: "BOTH" },
+          { title: "Desktop only", value: "DESKTOP" },
+          { title: "Mobile only", value: "MOBILE" },
+        ],
+      },
+    },
+  ],
+  preview: {
+    select: {
+      text: "text",
+      visibility: "visibility",
+    },
+    prepare: ({ text, visibility }) => {
+      if (!text?.length) return { title: "" }
+
+      const child = text[0].children
+
+      if (!child?.length) return { title: "" }
+
+      let title = ""
+
+      child.map((c) => (title += ` ${c?.text}`))
+
+      const subtitle = visibility.charAt(0) + visibility.slice(1).toLowerCase()
+
+      return {
+        title,
+        subtitle,
+      }
+    },
+  },
+}
