+++ cms/src/schemas/project/projectBlocks.ts
@@ -1,0 +1,22 @@
+export default [
+  {
+    group: "blocks",
+    name: "blocks",
+    title: "Blocks",
+    type: "array",
+    of: [
+      {
+        title: "Text block",
+        type: "projectBlockRichText",
+      },
+      {
+        title: "Media block",
+        type: "projectBlockMedia",
+      },
+      {
+        title: "Media block (full width)",
+        type: "media",
+      },
+    ],
+  },
+]
