+++ cms/src/schemas/schema.ts
@@ -2,12 +2,19 @@
 import createSchema from "part:@sanity/base/schema-creator"
 import schemaTypes from "all:part:@sanity/base/schema-type"
 
-// Custom types
-import blockRichText from "./custom/blockRichText"
-import media from "./custom/media"
-import mediaImage from "./custom/mediaImage"
-import mediaVideo from "./custom/mediaVideo"
-import year from "./custom/year"
+// Objects
+import blockRichText from "./objects/blockRichText"
+import media from "./objects/media"
+import mediaImage from "./objects/mediaImage"
+import mediaVideo from "./objects/mediaVideo"
+import projectAward from "./objects/projectAward"
+import projectBlockMedia from "./objects/projectBlockMedia"
+import projectBlockRichText from "./objects/projectBlockRichText"
+import projectRole from "./objects/projectRole"
+import projectThumbnail from "./objects/projectThumbnail"
+import settingsCatchphrases from "./objects/settingsCatchphrases"
+import settingsLiteral from "./objects/settingsLiteral"
+import settingsThemeColor from "./objects/settingsThemeColor"
 
 // Schemas
 import project from "./project"
@@ -17,12 +24,19 @@ import sticker from "./sticker"
 export default createSchema({
   name: "default",
   types: schemaTypes.concat([
-    // Types
+    // Objects
     blockRichText,
     media,
     mediaImage,
     mediaVideo,
-    year,
+    projectAward,
+    projectBlockMedia,
+    projectBlockRichText,
+    projectRole,
+    projectThumbnail,
+    settingsCatchphrases,
+    settingsLiteral,
+    settingsThemeColor,
 
     // Schemas
     project,
