--- cms/src/schemas/settings/settings.theme.colors.ts
@@ -1,54 +1,0 @@
-export default [
-  {
-    group: "theme",
-    name: "colors",
-    title: "Theme colors",
-    type: "array",
-    options: {
-      sortable: false,
-    },
-    of: [
-      {
-        name: "color",
-        title: "Color",
-        type: "object",
-        fields: [
-          {
-            name: "background",
-            title: "Background",
-            description: "Used as background color on the homepage",
-            type: "color",
-            validation: (Rule) => Rule.required(),
-            options: {
-              disableAlpha: true,
-            },
-          },
-          {
-            name: "accent",
-            title: "Accent",
-            description: "Used as text color on the homepage",
-            type: "color",
-            validation: (Rule) => Rule.required(),
-            options: {
-              disableAlpha: true,
-            },
-          },
-        ],
-        preview: {
-          select: {
-            background: "background",
-            accent: "accent",
-          },
-          prepare: ({ background, accent }) => {
-            if (!background || !accent) return { title: "" }
-
-            background = background.hex.toUpperCase()
-            accent = accent.hex.toUpperCase()
-
-            return { title: `${background} + ${accent}` }
-          },
-        },
-      },
-    ],
-  },
-]
