+++ cms/src/schemas/settings/settingsLiterals.ts
@@ -1,0 +1,11 @@
+export default [
+  {
+    group: "literals",
+    name: "literals",
+    title: "Literals",
+    description:
+      "These are the texts that populate all the website: buttons, labels, etc.",
+    type: "array",
+    of: [{ type: "settingsLiteral" }],
+  },
+]
